requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.0
sentence-transformers>=3.0.0
numpy>=1.24.0
torch>=2.0.0
tqdm>=4.60.0
